+++ backend/canisters/user_index/api/src/updates/start_user_migration.rs
@@ -1,0 +1,23 @@
+use oc_error_codes::OCError;
+use serde::{Deserialize, Serialize};
+use serde_bytes::ByteBuf;
+use types::{BuildVersion, CanisterId, UserId};
+
+// Only available in test mode, until the UserIndex migrates users itself
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub user_id: UserId,
+    pub multi_user_canister_id: CanisterId,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success(SuccessResult),
+    Error(OCError),
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct SuccessResult {
+    pub user: ByteBuf,
+    pub wasm_version: BuildVersion,
+}