+++ backend/canisters/user_index/impl/src/updates/mod.rs
@@ -53,6 +53,7 @@ pub mod set_user_upgrade_concurrency;
 pub mod set_username;
 mod set_vault_legal_hold;
 pub(crate) mod set_vault_reviewers;
+mod start_user_migration;
 pub mod submit_proof_of_unique_personhood;
 pub mod suspend_user;
 pub mod unsuspend_user;